+++ src/composables/useMarkdownActions.ts
@@ -1,6 +1,19 @@
 import { reactive, nextTick } from 'vue'
 import type { Ref } from 'vue'
-import { InlineFormatWrapper, DEFAULT_CODEBLOCK_LANGUAGE, MARKDOWN_TEMPLATE_CODEBLOCK, MARKDOWN_TEMPLATE_TASK, MARKDOWN_TEMPLATE_TASK_COMPLETED, MARKDOWN_TEMPLATE_UL, MARKDOWN_TEMPLATE_OL, MARKDOWN_TEMPLATE_BLOCKQUOTE, MARKDOWN_TEMPLATE_TABLE, NEW_LINE_CHARACTER, MARKDOWN_TEMPLATE_LINK } from '@/constants'
+import {
+  InlineFormatWrapper,
+  DEFAULT_CODEBLOCK_LANGUAGE,
+  MARKDOWN_TEMPLATE_CODEBLOCK,
+  MARKDOWN_TEMPLATE_TASK,
+  MARKDOWN_TEMPLATE_TASK_COMPLETED,
+  MARKDOWN_TEMPLATE_UL,
+  MARKDOWN_TEMPLATE_OL,
+  MARKDOWN_TEMPLATE_BLOCKQUOTE,
+  MARKDOWN_TEMPLATE_TABLE,
+  NEW_LINE_CHARACTER,
+  MARKDOWN_TEMPLATE_LINK,
+  MARKDOWN_TEMPLATE_IMAGE,
+} from '@/constants'
 import type { InlineFormat, MarkdownTemplate } from '@/types'
 
 /**
@@ -401,6 +414,15 @@ export default function useMarkdownActions(
   }
 
   /**
+   * Highlights text in a textarea on a given position.
+   * @returns {void}
+   */
+  const selectText = (start: number, end: number): void => {
+    const textarea = getTextarea() as HTMLTextAreaElement
+    textarea.setSelectionRange(start, end)
+  }
+
+  /**
    * Insert a markdown link at the current cursor position.
    * @returns {Promise<void>}
    */
@@ -419,6 +441,8 @@ export default function useMarkdownActions(
       const startText = rawMarkdown.value.substring(0, selectedText.start)
       const endText = rawMarkdown.value.substring(selectedText.end)
       let newContent: string = ''
+      const urlTextLength = 3
+      const textTextLength = 4
 
       // If text is selected, check the type of selected text and insert the link template around it
       if (selectedText.text.length !== 0) {
@@ -431,7 +455,7 @@ export default function useMarkdownActions(
         // Check if the selected text is a URL
         const isUrl = /^http(s)?:\/\//.test(selectedText.text)
         // Prepare the content
-        newContent = isUrl ? MARKDOWN_TEMPLATE_LINK.replace(/text/, '').replace(/url/, selectedText.text) : MARKDOWN_TEMPLATE_LINK.replace(/text/, selectedText.text)
+        newContent = isUrl ? MARKDOWN_TEMPLATE_LINK.replace(/url/, selectedText.text) : MARKDOWN_TEMPLATE_LINK.replace(/text/, selectedText.text)
 
         // Update the markdown
         rawMarkdown.value = startText + newContent + endText
@@ -446,6 +470,8 @@ export default function useMarkdownActions(
           textarea.selectionStart = startText.length + selectedText.text.length + 3
           textarea.selectionEnd = startText.length + selectedText.text.length + 6
         }
+
+        selectText(textarea.selectionStart, textarea.selectionStart + (isUrl ? textTextLength : urlTextLength))
       } else {
         // No text is selected
 
@@ -456,9 +482,9 @@ export default function useMarkdownActions(
         }
 
         // Prepare the content
-        newContent = MARKDOWN_TEMPLATE_LINK.replace(/text/, '')
+        newContent = MARKDOWN_TEMPLATE_LINK
 
-        let cursorPosition = 1
+        let cursorPosition = 7
 
         // Check if we need a space before or after the template
         if (/\w+$/.test(startText)) {
@@ -476,9 +502,100 @@ export default function useMarkdownActions(
 
         // Set the cursor position
         textarea.selectionEnd = selectedText.start + cursorPosition
+
+        selectText(textarea.selectionEnd, textarea.selectionEnd + urlTextLength)
       }
     } catch (err) {
       console.warn('insertLink', err)
+    }
+  }
+
+  /**
+   * Insert a markdown image at the current cursor position.
+   * @returns {Promise<void>}
+   */
+  const insertImage = async (): Promise<void> => {
+    try {
+      const textarea = getTextarea()
+
+      if (!textarea) {
+        return
+      }
+
+      // Update the selected text object
+      getTextSelection()
+
+      // Get the text before and after the cursor
+      const startText = rawMarkdown.value.substring(0, selectedText.start)
+      const endText = rawMarkdown.value.substring(selectedText.end)
+      let newContent: string = ''
+      const urlTextLength = 3
+      const altTextLength = 3
+
+      // If text is selected, check the type of selected text and insert the link template around it
+      if (selectedText.text.length !== 0) {
+        // If the user tries to click the button twice (with `url` selected) exit early
+        if (selectedText.text === 'url' && startText.endsWith('(') && endText.startsWith(')')) {
+          await focusTextarea()
+          return
+        }
+
+        // Check if the selected text is a URL
+        const isUrl = /^http(s)?:\/\//.test(selectedText.text)
+        // Prepare the content
+        newContent = isUrl ? MARKDOWN_TEMPLATE_IMAGE.replace(/url/, selectedText.text) : MARKDOWN_TEMPLATE_IMAGE.replace(/alt/, selectedText.text)
+
+
+        // Update the markdown
+        rawMarkdown.value = startText + newContent + endText
+
+        // Always focus back on the textarea
+        await focusTextarea()
+
+        // Set the cursor position
+        if (isUrl) {
+          textarea.selectionEnd = selectedText.start + 2
+        } else {
+          textarea.selectionStart = startText.length + selectedText.text.length + 4
+          textarea.selectionEnd = startText.length + selectedText.text.length + 7
+        }
+
+        selectText(textarea.selectionStart, textarea.selectionStart + altTextLength)
+      } else {
+        // No text is selected
+
+        // If the user tries to click the button twice (with the cursor in between the brackets) exit early
+        if (startText.endsWith(MARKDOWN_TEMPLATE_IMAGE.split('alt')[0]) && /^\]\((.*)+\)/.test(endText)) {
+          await focusTextarea()
+          return
+        }
+
+        // Prepare the content
+        newContent = MARKDOWN_TEMPLATE_IMAGE
+
+        let cursorPosition = 7
+
+        // Check if we need a space before or after the template
+        if (/\w+$/.test(startText)) {
+          newContent = ' ' + newContent
+          cursorPosition++
+        } else if (/^\w+/.test(endText)) {
+          newContent += ' '
+        }
+
+        // Update the markdown
+        rawMarkdown.value = startText + newContent + endText
+
+        // Always focus back on the textarea
+        await focusTextarea()
+
+        // Set the cursor position
+        textarea.selectionEnd = selectedText.start + cursorPosition
+
+        selectText(textarea.selectionEnd, textarea.selectionEnd + urlTextLength)
+      }
+    } catch (err) {
+      console.warn('insertImage', err)
     }
   }
 
@@ -597,5 +714,6 @@ export default function useMarkdownActions(
     insertMarkdownTemplate,
     insertLink,
     insertNewLine,
+    insertImage,
   }
 }