+++ vitest.config.ts
@@ -1,7 +1,6 @@
 import { defineConfig, mergeConfig } from 'vitest/config'
 import viteConfig from './vite.config'
 
-// @ts-ignore
 export default mergeConfig(viteConfig, defineConfig({
   test: {
     globals: true,